feat: Complete Flutter food delivery app with full architecture
- Implement comprehensive food delivery application with Flutter/Dart
- Add complete Bloc architecture with proper feature separation
- Include restaurant browsing, order tracking, chat, call features
- Add user profile management and onboarding screens
- Implement real-time database integration with Supabase
- Add comprehensive testing and error handling
- Support push notifications and image uploads
- Add category browsing with filtering and search capabilities
- Remove sensitive credentials from source code (use placeholders)

+++ food_delivery_app/lib/core/services/performance_service.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:collection';
-import 'dart:math' as math;
 import 'package:flutter/foundation.dart';
 import '../services/logger_service.dart';
 
@@ -129,8 +128,8 @@ class PerformanceService {
 
     return PerformanceStats(
       count: count,
-      min: min,
-      max: max,
+      min: min.toInt(),
+      max: max.toInt(),
       mean: mean,
       median: median,
       p90: p90,
